refactor(ATT-381): centralize compose spec helpers into shared test util
Address Sourcery review: extractServiceBlock duplicated across generator
specs. Move ROOT/readFile/fileExists/extractServiceBlock into
compose-test-utils.ts and import from technitium-dns + config-ui specs.
